+++ NEWS.md
@@ -4,7 +4,7 @@ Changes worth mentioning.
 
 ---
 ## 0.17.4 - 2017-09-04
-- [Feature] Added a new pattern generator
+- [Feature] New pattern rails generator using a set of default templates
 
 ## 0.17.3 - 2017-08-16
 - [Improvement] Upgrade gem with changes from [mountain view 0.12.2](https://github.com/devnacho/mountain_view)